dct/requirements/common.txt
celery==5.3.6
django==5.0.3
django-storages==1.14.2
django-redis==5.4.0
django-robots==6.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
flower==2.0.1
Pillow==10.2.0
psycopg==3.1.18
psycopg2-binary==2.9.10
requests==2.31.0
pytest==8.0.0
pytest-django==4.8.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.5.0 